Fixed: JSON passing to template was wrong
Twig escapes all strings containing quotes, changing each " by &quote;
which breaks JSON. This is fixed with |raw filter.

In addition, there were delimiters around the value of the array, which
caused the value to be a string instead of an array, which made
impossible to access its elements from javascript (unless the string
were parsed as JSON again, but this is pointless).