First Angular experiment
It shows a list of problems. This list is static, directly assigned from
controller. Next version has to download it from the server, via $http

The list can be filtered by query. Each problem provides links to
view/edit, which still are not functional (they should go to a different
view)