Preliminar version of angular with views/templates
There are two partial views (in `partials`) for different "screens" of
the application. The dependency of the new module is declared in
`app.js`. It requires that the main html file loads
`javascript/angular-route.js` first.

The list view shows a list of all the problems. The edit view (TODO)
shows details of single problem.