Fixed: drag&drop on filtered lists
Using angular filters to show only part of a list breaks the drag&drog
functionality, because if you drag the element N of the filtered list,
the inserted element will be the element N of the non-filtered list
instead.

The fix involves:

* Not using angular filters, but instead ng-show="function()" to
  determine via a function if the element is visible or not. This way,
  if the function returns false, the element is not inserted and it
  does not "consume" an index
* Creating a filtered list each time the query changes (via ng-change)
  for efficiency. Other solutions in internet like the one in
http://stackoverflow.com/questions/19072689/angularui-correctly-updating-the-models-between-two-lists-with-a-filter-applied
  do not use this approach, but instead it builds a new filtered list
each time ng-show is evaluated (i.e: one time per item in the list)
which is much more inneficient.

Still, since the query changes as the user types, each new char implies
a new filtered list to be built. If the original list is very large,
this can be innefficient. Debouncing should be used.