Fixed: dragging to filtered list
When dragging from the "document list" to the "database list", if the
database was filtered some problems could happen. In particular, if a
few problems are selected by the filter and they are all dragged to the
document, the "source empty" condition was not detected. Also, if a
question is dragged from the document to a filtered list, the list of
items selected by the filter should be updated