Added responsiveness to navbar for small screens
Now, when the browser window is small, the navbar disappears and a "menu
button" appears to the right. Clicking that button makes the menus
reappear in a vertical list, instead of horizontally.

In addition, provision for an "AppName" always visible (which is a link
to the home screen) was added to the navbar