TeX downloading working
There are two scripts related to tex generation:

* `generar_examen.php` generates a temporal folder, stores the
  required contents, and creates a zip from it (deleting the
  temporal folder and keeping only the zip). It does not
  send the zip to the browser, however, but instead sends JSON data
  with the status of the operation (success or error), and a URL
  which has to be used to download the zip
* `get_zip.php` is part of the url returned by the above script. It
  receives (via GET) the absolute path to the required zip, and after
  checking it existence and validity, it sends it to the browser
  for downloading. After that, the zip is removed so it should be
  regenerated if more downloads are required.

The angular controller calls the first one sending the appropiate
parameters via POST and receives as answer the URL from which the zip
has to be downloaded. Then, it simply sets `window.location.href` to
that URL which causes the browser to start the download.