GUI improvements in lists
Problems list, documents list, and lists of problems in doc-details are
now limited in height and scrollable. This helps to keep accesible the
search query at the top and the "new" button at the bottom, which
makes the GUI more useable when the number of items to show is large

Also the indentantion of html code was tidied, to help the detection
of nesting problems or divs not properly closed.