Templating features for authors
As an experiment, the contents of the "instrucciones" field of the
document are processed as another template. This means the author can
put Twig expressions inside the "instrucciones" field.

The render passes as context for that rendering an array with variables
`num_preguntas` and `puntos`, so that the author can write in the
instructions things like: "This exam has {{num_preguntas}} preguntas",
etc. Also, this gives the author a place where he can insert "comments"
that will not appear in the final tex (the would be included as Twig
comments, i.e: {# Like this #})

The file `ejemplodb.sql` was modified to include an example of these
features (for the document dated 2011-01-27)

This idea could be used too for other fields in the database, such as
the text of questions and answers. This way the author could write
questions which automatically compute the answer using Twig
expressions, as for example (a trivial one):

* In the question:
  {% set number=20 %}
  What is the remainder of {{number}} divided by 3?

* In the answer:
  The remainder is {{number % 3}}

The idea is that the questions, answers and explanations of each problem
share the same context, so that they can "share" variables this way.