Doc-details: new buttons to add/remove problems
In addition of drag&drop and double-clic, the user can also add problems
to the current document clicking in the small "+" next to each problem,
and remove them from the document clicking in the small "-".

This is specially targeted to smartphones and tablets, where drag&drop
didn't work, and double-tap caused zoom.