Tags input autocompletion
I don't  like how it is implemented in this library. It shows a list
with **all** tags, and higlights those which match the typed one. I
would prefer the list to show only which match, but it is not easy to
perform that filtering, unless it is send as query to the server.

I'll try with a different library in a next revision