Better autocomplete
Now the list of suggested autocompletions does not contain the full list
of tags, but only those who matches with the current input. The user can
select one of this list using mouse or up/down keys and add it to the
list by pressing enter.

The implementation requires in the controller to use $q to have defers
and promises, and $filter to filter the list of all tags based on what
was typed.